BIG error in SNOW_simple_seb_crocus corrected
When SNOW is a CHILD, we did not account for short-wave radiation (!) which caused too cold temperatures also in the ground during spring. In get_derivatives_prognostic_CHILD, I have now added the missing term snow.TEMP.d_energy_seb , which represents S_in-S_out. Radiation penetration to the ground surface is still not accounted for, but this should not be a big problem.

Sebastian